square totally working completely ready to go. edited integer to .to_i ->>> .to_f

+++ app/controllers/application_controller.rb
@@ -6,7 +6,7 @@ class ApplicationController < ActionController::Base
   def calculate_square
     # prams = {"elephant"=>"42"}
 
-    @num = params.fetch("elephant").to_i
+    @num = params.fetch("elephant").to_f
     @square_of_num = @num ** 2
     render({ :template => "calculation_template/square_results.html.erb" })
   end
